Only mark a day as missing if there are less than the requested regions.
Before this commit, we would mark any day with an amount of regions not equal to the monitor as missing.

However the API can return 4 regions, while the monitor has 2 regions.
We would mark this as missing days.

This commit fixes that issue. We only consider a day as missing, if the amount of regions is less than the monitor.